+++ LanPlaylistServer/src/main/java/net/kokkeli/resources/models/ModelUser.java
@@ -11,12 +11,12 @@ import net.kokkeli.resources.Field;
 public final class ModelUser extends ViewModel {
     private final String username;
     private final Role role;
-    private final int id;
+    private final long id;
     
     /**
      * Creates user model
      */
-    public ModelUser(int id, String username, Role role){
+    public ModelUser(long id, String username, Role role){
         this.username = username;
         this.role = role;
         this.id = id;
@@ -45,7 +45,7 @@ public final class ModelUser extends ViewModel {
      * @return Id of user
      */
     @Field
-    public int getId(){
+    public long getId(){
         return id;
     }
 }
